Refactored notification to allow for command success/fail notifications to be propogated outside of the domain.

+++ TheLongRun-League-Function/Projections/LeagueSummaryInformationProjection.cs
@@ -149,7 +149,7 @@ namespace TheLongRunLeaguesFunction.Projections
 
                             return new ProjectionResultsRecord<Get_League_Summary_Definition_Return>() {
                                 CurrentAsOfDate= prjLeagueInfo.CurrentAsOfDate,
-                                CurrentSequenceNumber = prjLeagueInfo.CurrentSequenceNumber ,
+                                CurrentSequenceNumber = (int)prjLeagueInfo.CurrentSequenceNumber ,
                                 AggregateTypeName = leagueEvents.AggregateTypeName ,
                                 DomainName = leagueEvents.DomainName,
                                 EntityUniqueIdentifier = leagueEvents.AggregateInstanceKey ,
